incremental checkin
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@2831 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/rxapi/rexxapi/server/RegistrationManager.hpp
@@ -85,7 +85,7 @@ public:
     const char *name;                  // the registered name
     const char *moduleName;            // module name for object (optional)
     const char *procedureName;         // procedure name (optional)
-    uint32_t userData[2];              // user area information
+    uintptr_t userData[2];             // user area information
     uintptr_t entryPoint;              // target entry point
     int     callType;                  // indicates the style of interface
     size_t  dropAuthority;             // Permission to drop
